[PORT] pingen to 9.0 - Reactivate model - move files to folder models, views, data - use name= instead of string= in views
